chore(tools): add copyright checker
Add a copyright check script under 'tools'. This script should be able
to handle different styles of comment prefixes (e.g. '//', '#'), and
skip the shebang line (e.g. '#!/bin/bash') if it appears at the top of
the file.

Add 'make presubmit' target, which is meant to be run as the presubmit
test.

Change-Id: Id0d92fd995cde60f05408c05501dc50931004369