feat(fixtures): adds initial fixture package
The package fixtures exposes a class and some utilities that can to
dynamically create objects required for UI mocks and tests. New fixture
generators can be added to the this package as needed.

See the API docs or tests for usage instructions.

Fixtures added in this CL include:

* Sequences: Create and retrieve namespaced sequences to generate
  incrementing values like "user-{seq}@example.org".
* Names: Generates unique, random names derived from US Census data.
* Users: Generates random users.

This CL also adds a small edit to the User model requiring constructor
arguments for name & email.

SO-11 #done

Planned follow up work to this is:

* SO-28 Add fixtures to tests and gallery.
* SO-29 Add a fixture for generating avatars.
* SO-30 Refactor dir structure and exports.

Change-Id: I0e567bb313959ffbb0c514c4b50959b6b6532972