feat(secure_store): add a simple mock API for the secure store service
The first .mojom based mock API for a secure store service. For now,
it is assumed that the store is already scoped to the current user and
the specific module that is using this service.

It is a really simple string key/value store, but we could add more
features in the future.

SO-59 #comment initial version

Change-Id: I566f4537cb839b8c150ee500198658bb3194e5e5