fix: prevent bound proxy objects from being garbage collected
When a proxy object is bound to some handle, someone should be holding
that reference so that the proxy doesn't get garbage collected.

The specific anti-pattern to avoid is to create a proxy as a local
variable and bind it to an interface handle.

So, here's what I did overall.
* When not using a given handle, don't even bind it.
* When some proxy object is actually needed, bind the handle and keep
  the proxy object as an object field.
* Call ProxyController::close() method in Module::stop().

Change-Id: I14a47ba5b048fd568e02d4bb2bd787f2e64f417b