[suggestion_engine] Fidl for ask-enabled agents
Originally for 11/30 we were just going to publish Ask queries to context and
have a Next agent surface these, which will likely end up being the case for
adapting Next-only agents to Ask. However, it turned out doing so was
equivalent in complexity with adding a straw-man Ask-wired suggestion agent
interface.

Change-Id: I26f4b58dc151820884086ae3c074e29ba88a7820