chore(make): backup symlinked .packages while running tests
'flutter test' tool internally runs 'flutter packages get' to update
the dependencies necessary to run the flutter tests. However, when
there is an existing .packages file which is a symlink, flutter test
overwrites the contents of the .packages file, which can break the
fuchsia build in a bad way.

As a workaround, the symlinked .packages files are now backed up
before running tests, and then restored.

Change-Id: I1e4a41dfe3ef4d22a34ad34fb6bc54e482b25a5f