Add Android code & docs
Adds instructions for building an Android compatible xi-core binary,
copying it to a connected device, and running the Flutter version
of the app.

A wrapper for communication with the xi-core binary/service has been
added to normalize a single interface to the xi-core client API and
help mitigage platform differences between Android and Fuchsia.

Change-Id: I4db5068e1445c77635870455f9ee4f9c27198af3