feat(widget_specs): draw the actual widget along with its description
Draw the actual widget in the gallery screen.

To make this happen, the extractor now also extracts the package name
and the relative dart file path in which the widget class is defined.

For now, pretty much only the AdditionalItems widget can be correctly
drawn, because it does not require any particular parameters.

TODO: generate and provide required parameters so that the widgets can
      be reasonably displayed.

SO-56 #comment draw the actual widget

Change-Id: I0539de188f197949da991c3a09be1bbb4b3c9853