feat(widget_specs): add @ExampleValue annotation
This CL adds a new package called 'widgets_meta' which defines the
@ExampleValue annotation that can be used in front of a constructor
parameter of a widget. For example,

    class Foo extends StatelessWidget {
		final int bar;
		final int baz;

		Foo({
			@ExampleValue(42) @required this.bar,
			this.baz,
		}) : super(key: key);

		// ...
	}

This value can be retrieved by the widget specs extractor, and this is
demonstrated in the extract_test.dart.

TODO: Use this example value as the initial value of that parameter
      when generating the widget page in gallery.

SO-171 #comment add ExampleValue annotation

Change-Id: Ib8da502584b524c014ccd9143849addb05ff2204