feat(widget_specs): provide default callback implementations
Some widgets expects callback functions to be provided, and those
callback functions are '@required'. Thus, in order to display these
widgets properly, the widget spec generator must provide a default
implementation for these callback parameters.

This CL adds the ability to provide those default implementations,
which just prints out a log indicating that the callbak is called with
the list of parameters provided to the callback.

SO-172 #done

Change-Id: I63ed3def2f3fcb3cdb3e900c45a93dc383c77073