Starting to develop xi front-end
Work in progress. So far, this patch establishes json-rpc
communication with the core. Also, a bit of cleanup to make it build,
initialize, and run on both Fuchsia and Android.

A new "Editor" flutter stateful widget maintains a line cache,
which is updated according to the new xi update protocol.

Change-Id: Ied2d6c1a782512883382eb4f4421d36cdf749770