feat(content_provider): read conversation list / message history from Ledger
Now the chat content provider agent obtains the Ledger instance and
passes it to the ChatContentProviderImpl, which will be manupilating
the chat history data in Ledger.

Right now, two of the FIDL methods, GetConversations() and
GetMessageHistory(), are now reading data from the Ledger pages.

To test this working, the agent is adding some random data to the
Ledger when it starts up. This code will go away and will be replaced
by some tests using the modular TestRunner service.

SO-329 #done

Change-Id: Ifad55afaf97ffa50dfffadd7f68071ee919af481