feat(content_provider): improve error handling
This improves the overall error handling in the chat content provider.
The general motivation is to make the agent implementation as stable
and resilient to errors as possible, because it's meant to be run as a
singleton service for all Chat modules. At a high level, the changes
are as follows.

* ChatContentProvider now returns a ChatStatus code along with the
  actual results. For each point of potential failure, the code uses
  try-catch block to catch the error and return the relevant status
  code as the result of the FIDL method call.

* For each content provider method, wrap the entire method body with a
  try-catch block, just in case some unanticipated error occurs in the
  logic. In this case, the method will log the error to the console
  and return ChatStatus::UNKNOWN_ERROR code.

* Use try-finally blocks to make sure that local proxy channels are
  closed properly even when some error occurs.

* Avoid throwing exceptions within a callback method body for two
  reasons. (1) The caller would not be expecting an exception, and (2)
  throwing an exception there would not be caught in a safe way. So
  instead, Completer<T> is used extensively to handle any errors
  outside the callback method bodies.

Also removes the User FIDL struct, because it's not really used for
now and unnecessarily making things more complicated. We should later
integrate this with the Contacts agent.

SO-331 #done

Change-Id: I7429bb2c4f2a4e5c0c7e1aadf046434994afc2ae