[cleanup] Start breaking out context_repository tests
This is the very first step of breaking out testing of the functionality
of ContextRepository into its own gunit-style unit test.

The goal is to create a shearing layer between async FIDL behavior and
policy with ContextRepository to make it easier to test. The integration
tests for context_engine will eventually only test that the FIDL
connections are made properly.

Change-Id: Ieb4de00c9a19d55f39265d95eb805383dfab3d3d