[context] Aggregate focal entities for each Story.
* Writes any topics coming from a specific Module to a Module-scoped
  part of the topic namespace.
* Introduces ContextCoprocessor: a lightweight interface for performing
  processing on Context topics in the scope of a single "set" operation.
  Long term we may want to expand our query language to have the
  coprocessors subscribe the way any other client would, but for now
  they are simple and completely contained within the implementation of
  the context_engine.
* Creates one coprocessor which looks for changes to a specified topic in
  one Module in a story, and aggregates values across all Modules into
  one Story-scoped list.
* Add an aggregator for "focal_entities" to the ContextEngineImpl.

A future CL will create another coprocessor which copies all values in
the "focused" story to the "/story/focused" topic namespace. Although
this is duplication of data (for now), it can be implemented easily.

Awkwardness exposed here that we should consider changing:

* The internal representation of topics as simple strings is inefficient
when performing operations that deal with the scoped namespace nature of
the semantics of the string. I've left it in for now since, again, it's
simple, but the implementation is inefficient since translations are
constantly necessary back and forth between the string form and an
internal representation that exposes the scopes.
* We might consider storing the actual values in ContextRepository as
rapidjson values instead of strings. This would make manipulations
easier.

Change-Id: Idda84566bd7c22fc7f1f4545632aa67e0d45cc1f