Width Constraints Based Layout Algorithm
Calculates a layout by dividing the screen horizontally based on
copresentation requests and width constraints. For now the width
constraints globally restricted to the max of 300 pixels or 1/4 the
width of the screen.

Properly animates multiple widgets that are animating out, and orders
touched widgets on top ofthers.

Also prepares for the upcoming StoryShell API change by introducing
those concepts to the SurfaceGraph and layout widgets. These include
SurfaceProperties (intrinsic surface needs) and
SurfaceRelation (associations and affinities to other surfaces).

SY-161

Change-Id: I966434f2357ffba9da51aad23a50877a9d8f78b8