Keep Dismissed Surfaces In SurfaceGraph
To allow future resummoning of dismissed surfaces, we keep them in the
SurfaceGraph. Dismissed surfcaces are removed from the focus stack and
added to a blacklist to prevent display. When refocused it is removed
from the 'dismssed' blacklist.

The layout algorithm removes dismissed surfaces from the copresent tree,
and collapses their children to the parent node. Dimsissed nodes are
still included in emphasis calculations to keep surface sizes
consistent.

SY-165 #done

Change-Id: If6b11e3d207f3d8809b35009702c58784f9604c1