Implement FocusView
FocusView removes a view from the dimissed list and adds it to the top
of the focus stack, causing a layout. Changes  absoluteEmphasis
calculation to account for a relative surface not being a direct parent
of both surfaces.

Also adds UI in example module to focus child modules.

SY-164 #done

Change-Id: I482d7126f5fb5becc1f0edb72c7249f114a64427