Add script and build rules to download and use prebuilt libwebkit.so
This switched the default configuration to using prebuilt binaries for
libwebkit.so since that target is excessively heavy to build in the
normal build. To override and build locally, pass the gn argument
'use_prebuilt_webkit=false'. The download script puts prebuilt shared
libraries into prebuilt/{x86_64,aarch64}/libwebkit.so along with a
stamp file, identified by the git commit hash of the webkit revision
the library was built from.

Change-Id: I8ecad449fb1171d55469ffffe58af52fc5dbb6e1