Add module_url to AddModuleToStory suggestion proposal.
This untangles use of module_id vs module_url which was previously
overloaded.

More practically, it makes it possible to add multiple module instances
of the same module url to a story. We update action log for the video
example to use 'video_id' as the module_id. This makes it possible to
start multiple 'share video' presenter screens at the same time.

Change-Id: If5a89baf78a3ae75567accdfbb5258c199e1f4b5