Focus Relative to Closest Ancestor that Has Focus
To prevent having visible surfaces have no power to focus their
children, we ensure that if the relativeId or its children are not in
the focus stack, we find the closest ancestor that is in the focus stack
and insert after that. If no ancestor is found in the focus stack, then
it will not be focused.

Change-Id: Ic96fb70e83a63ef24ba40bf5f73c58ff5a79d8f6