Track mx handles per-isolate.
This just just the //lib/tonic part of the change. There are also
changes in //flutter, //apps/dart_content_handler and //lib/fidl.

As handles are passed from C++ to Dart they are added to a per-isolate
set and circular shifted 16 bits. As they are passed from Dart to C++
they are circular shifted 16 bits and checked against that set to make
sure the belong. Handles can be closed or removed from an isolate (to
support mx_channel_write for example). When the isolate shuts down all
handles in the set are closed.

Change-Id: I506fb2d53941ddb246b546979dbb079217398629