feat(agent): waiting for the conversation information
When the conversation module gets the conversation id from the Link
"before" the conversation meta information is synced in the chat
Ledger, the conversation module might fail to get the conversation
title properly.

In this CL, newConversation() method now accepts a boolean "wait"
parameter, which allows waiting for the specified conversation to
appear in the conversations ledger page. This solves the situation
where the newly created conversation title is not properly obtained
from another device logged in as the same user.

SO-566 #done
SO-596 #done

Change-Id: I45d5e9c18093713f9543f20ced1405964a5fb63e