[agents] Remove module_suggester.cc
The cxx version of module_suggester was a hold over from the Nov'16
demo push, and was no longer making good or valuable suggestions.
This CL removes it. Future module suggestions should be implemented
in the dart version of this module.  The YouTube and terminal
suggestions from the cxx suggester have been moved to dart.

MW-133 #done

Change-Id: Ic8939f2bfbd4c97918fabfcd8a72d1bccb5b297a