[context] Make ContextMetadata explicitly store values.
I was previously under the impression that Dart made it hard to work
with structs with optional values. I was wrong, and just had a lot of
examples using an anti-pattern. This CL makes each context metadata
field explicit.

I expect the number of fields to grow substantially over time, and
having a central place where they are explicitly documented is nice.

MW-137

Change-Id: I2ac1f327ed99434a824e8bd747d662466ebc52d6