refactor(agent): remove duplication of PageWatchers
The BasePageWatcher now keeps track of the page snapshots, so that the
last known snapshot of that page can be reused in chat content
provider implementation. With this change, the main page (which lists
all the conversations) only has one ConversationListWatcher associated
with it, and each conversation page can have at most one
ConversationWatcher.

SO-368 #done

Change-Id: I3a049ad9b1cf583bbf6769d78f948e58a1953b33