[usage_log agent] Initial implementation
This is a start on a usage logging agent that reads data on
focused modules from the context service and sends the URLs of
these modules to Cobalt for logging.  This may not be the right
long term approach, but it gives us a starting point for future
usage logging.  Also adds the cobalt service to the set of FIDL
services available to agents.

Change-Id: I7a9351109bc5304137494605bbeabb6223829868