refactor(models): break down ModuleModel
This CL separates a lot of the player logic from the casting/remote device
logic.

Previously, ALL model fields and logic were stored in the ModuleModel.
This meant that notifyListeners() was being called for all model changes,
even when the changes were not relevant to the particular widget.
All widgets subscribed to the single model. An example is when dragging
and dropping the Cast Thumbnail onto the Cast icon (one model event for
the Device Chooser UI), notifyListeners() was being called periodically
because the Scrubber UI was updating due to the MediaPlayerController's
progress field updating (this was resolved earlier with some conditionals).

In this CL, the ModuleModel is separated into the VideoModuleModel and the
PlayerModel. The VideoModuleModel is responsible for events that relate to
the device and remote devices. The PlayerModel controls anything to do
with video playback. This way, Casting/Device events should not affect
Widgets used for video playback (e.g. Scrubber, Play Controls), and vice
versa.

SO-542 #done

Change-Id: Ic87f84ce40201ae7ec62e23bc2f7721274c348cb