[context] Use metadata-based querying.
This is part of a large rewrite of the context engine. It is a departure
from the world of having one value per topic, and rather allows metadata
to be attached to a value (the topic is one metadata field) and then
allows querying across any combination of metadata fields.

MW-137 #comment

Change-Id: I7214f46cae3b115bf2c71dcd4e5d2d31f005ee4a