[context] Use metadata-based querying.
This is part of a large rewrite of the context engine. It is a departure
from the world of having one value per topic, and rather allows metadata
to be attached to a value (the topic is one metadata field) and then
allows querying across any combination of metadata fields.

MW-137 #comment

Change-Id: I27b1345edbad30abc1c2e76a8e9d4de550de76ce