[context] Use metadata-based querying.
This is part of a large rewrite of the context engine. It is a departure
from the world of having one value per topic, and rather allows metadata
to be attached to a value (the topic is one metadata field) and then
allows querying across any combination of metadata fields.

MW-137 #comment

Change-Id: I22848e2b058d535f87f034a46fc4e7a0033d0c0e