[context] Add ScopedContextValue library class.
This encapsulates the common mode of operation with the ContextWriter,
and allows the client to simply use "set state" operations on the value
instead of having to worry about if the value exists already (and thus
updating its state) or not (and having to add a new value).

Requring this class at all begs the question of if ContextWriter's
interface itself should change to expose something more like this class,
and I'm leaning towards "yes". This at least is available now for
clients to use, and will make it much easier to migrate them when/if the
ContextWriter interface changes.

Also:

* Add & implement ContextWriter.Remove()
* Use ScopedContextValue in the story_info acquirer.

Change-Id: I83227881d0edcb911f54b693a14dd7c7c56a2040