[maxwell] Suggestion Engine refactor
We are refactoring the suggestion engine and the ProposalPublisher
such that registering an AskHandler (now called a QueryHandler) will
not also require consumers to register a ProposalPublisher.  This CL
alters the UserShellImpl to pass through IntelligenceServices, which
provides access to registering both ProposalPublishers and
QueryHandlers rather than just passing through the ProposalPublisher
interface.

MW-233 #done

Change-Id: Ic972ee60ded774622163e98beadc337fb33aa884