[maxwell] Suggestion Engine refactor
We are refactoring the suggestion engine and the ProposalPublisher
such that registering an AskHandler (now called a QueryHandler) will
not also require consumers to register a ProposalPublisher.  This CL
alters the UserShellImpl to pass through IntelligenceServices, which
provides access to registering both ProposalPublishers and
QueryHandlers rather than just passing through the ProposalPublisher
interface.

Peridot now works, so this can go in.

MW-233 #done

This reverts commit ce541411baac60cf33820f27b04e1de9eaae3add.

Change-Id: I427ca69d54df55c12b29ceaf288338c394e2fe7c