[dart:zircon] Fix HandleWaiter destructor logic.
The Handle used to hold a ref to the HandleWaiter so that the waiter's
destructor wouldn't be called until after the handle's destructor. We
changed that didn't update the waiter's destructor to dissociate itself
from the handle if it's called before the handle's.

This is blocking some important stuff, apparently.

DNO-162 #done

Change-Id: I1e9baaa4fd29bc4c1f4bd4d82d07f7d7b3ef9814