Patching state transitions
Half of the state transitions migrated to the ask handler, but it
doesn't look like they were ever wired to the suggestion model. Fall
back on the SpeechStatus (which is still wired up, but will be
deprecated someday).

This fixes the issue where speech output cuts off due to prematurely
entering the listening-for-hotword state.

Change-Id: I8da4440f5f30d0c5c4b5fc8bedf693b416173117