fix(conversation): fix initialization race in conversation module
Since the LinkWatcher is registered before the initialization logic is
fully finished, the _fetchConversation() method could be called even
before the chat content provider service is requested. For this
reason, the getConversation() call was silently failing without
actually sending the FIDL message.

This CL fixes this problem by awaiting for the init logic to finish
before setting the conversation ID obtained from the Link.

SO-888 #done

Change-Id: I5d7e41263848c27019db46119f249420f7afd1e1