[armadillo] Updates timezone tracking.
The timezone offset (in minutes) is now no longer used to set the actual
clock, and is handled in the Dart VM when using the DateTime.now()
constructer for a time object.  Instead the timezone ID is used to track
which timezone has been selected in the timezone picker (the active
timezone is highlighted in the picker to show the current timezone).

Change-Id: I3019aa86ca96f032c07d7a2cc98d2e3b5663b595