[dart_runner] Create the fdio_ns_t in RunApplication
Previously, we created the fdio_ns_t in the DartApplicationController,
but once we switch to using Fuchsia packages for Dart apps, we'll get
the snapshot data from the fdio_ns_t, which means we need to create the
fdio_ns_t before we create the DartApplicationController.  This CL
prepares us for that change.

Change-Id: Ia931874b696a31611c6ffb903b8d764281ba7abb