[chat] Fail harder when the transport channel couldn't be initialized
The Firebase transport channel would never be successfully initialized
under the following conditions:

* When the necessary api keys are not provided
* When chat is launched in guest mode

Previously, even when this happened, the UIs looked normal, which gave
the wrong impression that everything is fine. With this CL, the chat
UI now shows a very prominent error UI instead of showing the normal
chat screen.

SO-1017 #done

Change-Id: Iacfa028203dad42ce60df9a919e27c50c90b82af