[chat] Extract Firebase DB connection logic into a separate agent
Refactored the Firebase DB related code in a way that the Firebase DB
client can be used for more general purpose other than just sending
and receiving chat messages.

With this CL, the firebase db connection is managed in a separate
agent, which can then be used by any other mods or agents.

The intention is to use the same Firebase
DB client for storing shared-data between users for embedded mods.

SO-1050 #comment Refactor Firebase DB client

Change-Id: If9b78ac474537c59e6c2049b219b0aca42802909