[chat] Share the sender and current user as well
Share the sender / current user email addresses to the embedded mods.
This is useful when the embedded mod wants to run some initialization
logic in only one place.

For example, if an embedded mod wants to pick a random number and
share the result with all the users in the conversation, the number
has to be picked only once. This can be done by the embedded mod by
running the random picking logic only if the sender is same as the
current user. The other users can just wait for the result to be
notified.

SO-1050 #done

Change-Id: I27a8ce1cb842d39d0cff9c5faf1f4a050da644aa