[web_view] Improved, dynamic entity extraction
This introduces a new, more extensible web entity extraction system.
It's written in Typescript for superior developer productivity but a
built and bundled version is checked in so that nobody needs to go and
install NPM if they just want to build Fuchsia.

The new entity extractor adds support for Microdata and listens for
mutation events so that entites can be re-scraped. When the entities
in the page change an event is triggered and the web view updates the
context store's list of entities.

Entity extraction is now gated behind a gn arg. To enable it pass:
`--args experimental_web_entity_extraction=true` to `fx set`.

MI-746 #done
MI-747 #done

Change-Id: Icf8e05ed2286663b826cfd57fc6816625b71b405