Relands: package:fuchsia and package:zircon
This change introduces two new packages living at
//topaz/public/dart/{fuchsia,zircon}. Their purpose
is to conditionally re-export dart:fuchsia and
dart:zircon when running on Fuchsia, and to
re-export mocks of those libraries otherwise.
This allows tests written against Dart's generated
FIDL bindings to run on the host. This approach
will also be applicable to FIDL2

Relands to fix release-mode build, and to
change "mocks" to "fakes" following comments.

fixes TO-697

Change-Id: I4cc0fa2b3986f92e96c37655b13a4da082edb97b