[auth] Adds impl for authorize endpoint.
This cl uses the web_runner interface to start a web_view that displays
OAuth login/authorization screens from gaia. At the end of the authorize
flow, a long lived refresh token is minted for the configured clientId
and scopes which is persisted by the Token Manager Service in Garnet.

AUTH-29 #done

Change-Id: I08269ae1dcb2791a1c14ef5542576a1e20bc7681