[mi_dashboard] Be more explicit in JSON generation
The MI dashboard has been failing to start up because it tries
converting some FIDL objects to JSON, some of which can't be converted.
This is a runtime breakage that can be caused by FIDL interface changes,
so they aren't picked up by build or automated tests. You don't even see
an error until a user logs in.

This change makes some of the JSON conversion more explicit. This should
break more eagerly when FIDL interfaces are changed. Generally it would
be good if the MI dashboard server was more explicit in which data it
wants to expose to the web frontend.

Disclaimer: I have only verified that mi_dashboard now starts, not that
the web UI actually works.

Change-Id: I7e9432917e7f8b7ebd99c2a0ae75e62dbb05d1b2